requirements.txt
azure-core==1.29.5
azure-data-tables==12.4.4
azure-functions==1.17.0
certifi==2023.7.22
charset-normalizer==3.3.2
exceptiongroup==1.2.0
idna==3.4
iniconfig==2.0.0
isodate==0.6.1
mock==5.1.0
multidict==6.0.4
packaging==23.2
pluggy==1.3.0
pytest==7.4.4
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
tomli==2.0.1
typing_extensions==4.8.0
urllib3==2.1.0
yarl==1.9.2
